use DCI instead of hardcoding the crud context
this makes the user and gift classes more flexible, as they are not
bount at the source code level to the crud context.

later on we will most likely be wrapping these entities with ActiveRecord
objects and we can create our own crud context at that point that deals
with ActiveRecord in particular.