move some methods into the postman class
i feel that the user class is trying to do too much at the moment. this can
be seen by it having to know how to manipulate the receiver object (it
is coupled by the method name of the receiver, even though it's actually
the same class).

the postman class now has the responsibility to make sure that the receiver
receives the package, the giver knows he's given the package, and the gift
knows from whom and to whom it was given.