refactor gift to use a repository
since i've started using forwardable during the refactoring, i realized i
can dry up a lot of the code for the user as well.

this seems to be an interesting pattern; entities would likely need to be
backed by a repository, and to make things easy i can provide a private
repository class as a default.

however i wonder if this actually belongs in the implementation? maybe it does
and it will give the users a good idea of how the repository is supposed to
be designed.

it may not as well, because the repository is not directly related to the
class. maybe i'll just need to provide a mock? we'll see as time goes.