move the crud context into the spec folder
i thought the crud context was needed by the classes, then i realized after
talking with jimmy that the context only is useful for tests. the context
was already decoupled so it was an easy task of just requiring the extracted
context into the point where the test is.

this has the added advantage of having all my classes not depending on
any of the other classes. i'm now able to remove the requires on the top
of each source file because they are decoupled enough that they just need
to rely on the (implicit) interface.