a little bit of refactoring to consolidate the things that do the same things
right now there seems to be a bit of something like feature envy. the gift
instance attributes are being manipulated by the postman, but putting that
feature into the gift class doesn't make a lot of sense.

gift seems to be a simple value object that contains the details about the
gift itself; having it contain other info makes it do more than one thing
and it shouldn't be responsible for that.

this makes it clear that a new object is required to handle the records associated
with gift transactions - maybe a TransactionRecord class? then all the other
objects will take on the TransactionRecord object and use that to query the
status of their gifts.

for example, the user.gifts_received method can optionally take on a custom
TransactionRecord object and calculate the value based on the entries of this
particular transaction record (similar to a notebook).