add the transaction record class
this class will contain the transaction records of gift giving. it functions
almost like a notebook, with the particulars of gifts being given and received,
the parties involved, and the context of the transaction (date, message, etc).

the other entities can then optionally use the transaction record (or in case
of tests, provide their own) and respond to queries such as:

* user.gifts_received # right now, the number of gifts received but later we
  can return an array of all gifts received by this user
* gift.given_count # something like a popularity count
* user.gifts_given # similar to gifts_received, but for gifts that have been
  given by this user